add sensors.camera_distance and don't do face rec on images while capturing...
The goal of not doing verification by running get names for faces on each frame during new face engagment is so that we can capture more frames and improve the recognition rate.  This was accomplished by shifting the validation of unknown single face to post processing and is instead done by retrain_model.py.